Array
Various exercises using arrays, the first exercise was to add up the
values in an array, the second was to then have an average for them, the
third one was to copy the array values into another one, the fourth one
is to sort them in my case, from the highest to the smallest. the last
one is not functional yet.